Reorganized code folders / refactored code
