+++ app/routes.h
@@ -1,0 +1,13 @@
+#ifndef APP_ROUTES_H_
+#define APP_ROUTES_H_
+
+namespace app {
+
+enum class Routes {
+    InitialScreen,
+    GameScreen
+};
+
+}  // namespace app
+
+#endif
