--- routes.h
@@ -1,11 +1,0 @@
-#ifndef TOWERDEFENSE_ROUTES_H_
-#define TOWERDEFENSE_ROUTES_H_
-
-// TODO: decide where to put this class
-
-enum class Routes {
-    InitialScreen,
-    GameScreen
-};
-
-#endif
