+++ view/hud/turretselectoritem.h
@@ -1,5 +1,5 @@
-#ifndef HUD_TURRETSELECTORITEM_H_
-#define HUD_TURRETSELECTORITEM_H_
+#ifndef VIEW_TURRETSELECTORITEM_H_
+#define VIEW_TURRETSELECTORITEM_H_
 
 #include <QGraphicsItem>
